Split out usePrevious hook.
